change upload static
