+++ package.json
@@ -19,6 +19,7 @@
     "express": "^4.17.1",
     "express-http-proxy": "^1.6.2",
     "expressjs": "^1.0.1",
+    "fs-extra": "^10.0.0",
     "joi": "^17.4.0",
     "jsonwebtoken": "^8.5.1",
     "lodash": "^4.17.21",
